Fix the problem with generate.sh
Seems like the confusion here was caused by a minor difference between
bash and sh (or different versions of sh?). Some versions apparently
deletes the contents the file that's the target of the cat heredoc
snippet, while others don't.

To be sure: *Do not write to the same file as the one to read from!*

Closes #9